Fix Parquet reader filter index
The filter column Id passed to DuckDB Parquet reader should be the
relative index to the projection list. If a filter column is not
included in the projection list, it is still included in the columnIds
array and the output from the reader will include this column. It will
be discarded in HiveConnector later when outputColumns is being built.