Unify updating ColumnReader readOffset_
Previously, the readOffset_ in the selective ColumnReaders is updated at
various places when the read is finished: 1) filterNulls();
2) readWithVisitor(); 3) StringColumnReader::readHelper(); 4) at the end
of read() itself for complex types. Among these functions, the first 3
were called by read() under different situations. There are also multiple
locations within a function that the readOffset_ is updated due to early
returns. All these made the debugging experience bad and it's hard to
find where the readOffset_ is updated.

We shall notice that all ColumnReaders need to update the readOffset_ at
the end of read(), and this commit unifies the updating of readOffset_
this way at one location.

The other reason of this change is because the filterNulls() and
readWithVisitor() may be called multiple times for a batch in the Parquet
reader, one for each page. However we may not want to advance readOffset_
until the whole batch finishes.

There is also a slight refactoring in some readCommon() functions to
remove the early return to make the debugging experience easier.