Move HiveDataSource out of HiveConnector files
The HiveConnector.cpp and HiveConnector.h are already very big. To
support Apache Iceberg, there will be more code added to HiveDataSource
and making these two files even larger. Note that the HiveDataSink class
is already in its own files. To prepare for the upcoming development,
We move the HiveDataSource class out of HiveConnector.cpp and .h to be
in its own files.