Refactor the operator timing code in Driver
Previously, a DeltaCpuWallTimer needs to be created at the beginning of
each operator call block. This part of code is usually greater than 4
lines, making the Driver::runInternal() function too long and hard to
read. This commit extracts the common code into withDeltaCpuWallTimer()
function, such that measuring the operator time is easy and clean.