Pushdown integer upcasts to source operators in LocalPlanner
Changed pushdown_integer_upcasts_to_scan to pushdown_integer_upcasts_to_source

PUshdown integer upcasts to source operators

Only when there are join nodes or result output nodes in the fragment
that an integer upcast is needed. In this case, pushdown the upcasts to
source operators like TableScanNode or ExchangeNode.